Add Rotary Encoder and Switch
Rotary Encoder appears to be mostly working with this software commit (though it seems like the internal state variable is sometimes allowed to progress beyond 128 leading to "going off the grid".
Rotary Potentiometer has been disabled due to a found hardware PCB routing bug. It will be re-added after fly-wire workaround is complete (The routing has already been fixed for next revision of the Core module PCB). Issue was that potentiometer was routed to a teensy pin that is unable to do an `analogRead`