Cleanup after all sensors working
Last commit was just to get all sensors validated to the point I knew that the next board spin was going to succeed. Now that the next revision of boards has been ordered, this step is to cleanup and normalize the code in preparation for generating Doxygen documentation. This is also probably the last commit for this PR